Event Listeners:
An event listener is added to the btn element. When clicked, it triggers the getWeather(city) function, which fetches weather data based on the input city.
Another event listener is added to the input element to listen for keydown events. When the "Enter" key is pressed, it also triggers the getWeather(city) function, allowing the user to submit the form by pressing "Enter" instead of clicking the button.