Made the header navigation bar:
If the heights of the child elements (search-container and category-container) are different, the items-center class will still vertically center them within the parent container based on the tallest element. In other words, it will align both child elements vertically so that their centers are at the same position, which will be determined by the tallest element.

For example, if the search-container is taller than the category-container, both elements will be aligned vertically so that their centers are at the same height, which is determined by the height of the search-container. Similarly, if the category-container is taller, then the alignment will be based on its height.

mx-auto equalizes the space before the container and after the container on both cides effectively centering the element