Made the product list:
Catch:

The border-width property specifies the width of the border itself, not the padding between the content and the border.

When you set a border width using CSS, it determines how thick the border line around the element will be. For example, if you set border-width: 2px;, the border will have a thickness of 2 pixels.

Padding, on the other hand, specifies the space between the content of an element and its border. It's controlled by the padding property in CSS.

increasing the border width will indeed make the border line thicker, but it won't affect the padding between the content and the border. If you want to adjust the space between the content and the border, you need to adjust the padding separately.

In Tailwind CSS, border-4 is a utility class that sets the border width to a size equivalent to 4px. It applies a border to the element with a width of 4 pixels.