script modules for ion icons:
Added the icons in the bottom of the screen usinf fixed because of which later we will need to use h-screen for background to cover the whole screen with that background since fixed and absolute:

The background may not be covering the whole section because of the fixed positioning applied to the container. When an element is positioned fixed, it is removed from the normal document flow and positioned relative to the viewport, which means it won't take up space in the layout like other elements.

<script type="module" src="https://unpkg.com/ionicons@7.1.0/dist/ionicons/ionicons.esm.js"></script>: This script tag imports the Ionicons library using ECMAScript Modules (ESM). It allows you to use Ionicons icons as custom elements (<ion-icon>) in your HTML and interact with them programmatically using modern JavaScript syntax.

<script nomodule src="https://unpkg.com/ionicons@7.1.0/dist/ionicons/ionicons.js"></script>: This script tag is a fallback for browsers that do not support ECMAScript Modules. It loads the Ionicons library using the traditional script tag syntax, allowing older browsers to still use the Ionicons icons.