Added banner class to the section in order to apply background according to light and dark mode:
Why h-screen?

The background may not be covering the whole section because of the fixed positioning applied to the container. When an element is positioned fixed, it is removed from the normal document flow and positioned relative to the viewport, which means it won't take up space in the layout like other elements.

or additionally remove the section and just put the banner class in body becoz then it will cover the whole screen