Toggle Classes:
banner.classList.toggle("bg-gray-900"): This toggles the class bg-gray-900 on the element with the class banner, changing its background color to gray when the dayNight element is clicked.
banner.classList.toggle("text-white"): This toggles the class text-white on the element with the class banner, changing its text color to white when the dayNight element is clicked.
day.classList.toggle("hidden"): This toggles the class hidden on the element with the class day, making it visible or hidden when the dayNight element is clicked.
night.classList.toggle("hidden"): This toggles the class hidden on the element with the class night, making it visible or hidden when the dayNight element is clicked.