Created the form and added tailwind css to it to make it visually appealing:
Used (label , for) tags:

for triggers an on click event on the label and transfers the focus to the tag specified by the id mentioned in the string with for.

In this example, the label "Username:" is associated with the input field for the username. The for attribute of the <label> element matches the id attribute of the input field, creating a link between them. This association improves accessibility and usability because clicking on the label will focus or activate the associated form control.

The type attribute is used in various form elements (<input>, <button>, etc.) to specify the type of control being created. For <input> elements, it defines the type of input data that the field will accept. Some common values for the type attribute include:

text: A single-line text input field.
password: A password input field (the entered text is masked).
email: An input field specifically for email addresses.
number: An input field for numerical values.
checkbox: A checkbox input.
radio: A radio button input.

To work with radio buttons in HTML, you need to create a group of radio buttons using the <input> element with the type="radio" attribute. Radio buttons allow users to select only one option from a set of options.

Here's how you can create a group of radio buttons:

<form>
  <input type="radio" id="option1" name="options" value="option1">
  <label for="option1">Option 1</label><br>

  <input type="radio" id="option2" name="options" value="option2">
  <label for="option2">Option 2</label><br>

  <input type="radio" id="option3" name="options" value="option3">
  <label for="option3">Option 3</label><br>
</form>

Explanation:

Each radio button is created using the <input> element with type="radio".
The name attribute is used to group the radio buttons together. All radio buttons with the same name attribute belong to the same group, and only one radio button in the group can be selected at a time.
The id attribute uniquely identifies each radio button.
The value attribute specifies the value that will be submitted if the radio button is selected.
The <label> elements are used to provide labels for the radio buttons. The for attribute of each <label> should match the id of its corresponding radio button, linking them together.
When a radio button is selected, its value will be submitted as part of the form data when the form is submitted. In JavaScript, you can use event listeners to detect when a radio button is selected or change its state programmatically.